backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.12.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.6.0
